refactor: unchecked itterations and stoarge variable udpate

+++ contracts/core/ReveelMain.sol
@@ -8,7 +8,7 @@ import "./RevenuePath.sol";
 
 contract ReveelMain is Ownable, Pausable {
     //@notice Fee percentage that will be applicable for additional tiers
-    uint256 private platformFee;
+    uint88 private platformFee;
     //@notice Addres of platform wallet to collect fees
     address private platformWallet;
     //@notice The list of revenue path contracts
@@ -31,7 +31,7 @@ contract ReveelMain is Ownable, Pausable {
     /** @notice Updates the platform fee percentage
      * @param newFeePercentage The new fee percentage
      */
-    event UpdatedPlatformFee(uint256 newFeePercentage);
+    event UpdatedPlatformFee(uint88 newFeePercentage);
 
     /** @notice Updates the platform fee collecting wallet
      * @param newWallet The new fee collecting wallet
@@ -53,7 +53,7 @@ contract ReveelMain is Ownable, Pausable {
 
     constructor(
         address _libraryAddress,
-        uint256 _platformFee,
+        uint88 _platformFee,
         address _platformWallet
     ) {
         if (_libraryAddress == address(0) || _platformWallet == address(0)) {
@@ -79,7 +79,7 @@ contract ReveelMain is Ownable, Pausable {
     /** @notice Set the platform fee percentage
      * @param newFeePercentage The new fee percentage
      */
-    function setPlatformFee(uint256 newFeePercentage) external onlyOwner {
+    function setPlatformFee(uint88 newFeePercentage) external onlyOwner {
         platformFee = newFeePercentage;
         emit UpdatedPlatformFee(platformFee);
     }
@@ -147,7 +147,7 @@ contract ReveelMain is Ownable, Pausable {
 
     /** @notice Gets the platform fee percentage
      */
-    function getPlatformFee() external view returns (uint256) {
+    function getPlatformFee() external view returns (uint88) {
         return platformFee;
     }
